git-svn-id: https://svn.code.sf.net/p/wired/code/trunk@1377 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ wired/src/engine/AudioEngine.h
@@ -1,4 +1,4 @@
-// Copyright (C) 2004-2006 by Wired Team
+// Copyright (C) 2004-2007 by Wired Team
 // Under the GNU General Public License Version 2, June 1991
 
 #ifndef __AUDIOENGINE_H__
@@ -145,7 +145,7 @@ static int	AudioCallback(const void *input,
 
 
   int nchan = 0;
-  vector<long>::iterator chan;
+  vector<int>::iterator chan;
 
   if (data->SampleFormat & paFloat32)
     {
